+++ pages/src/localdev.ts
@@ -5,11 +5,10 @@ import '../../src/polyfills/polyfills.es6';
 // Validate environment
 import '../../src/polyfills/polyfills.validate';
 
-import './common/test-media';
-import './common/test-media-source';
-
 // With Namespace
 import '../../src/modules/lib';
+// Config
+import './common/breakpoints';
 
 import {
   ESLImage,
@@ -44,6 +43,21 @@ import '../../src/modules/esl-media/providers/html5/video-provider';
 import '../../src/modules/esl-media/providers/youtube-provider';
 import '../../src/modules/esl-media/providers/brightcove-provider';
 
+import './esl-media-demo/test-media';
+import './esl-media-demo/test-media-source';
+
+import {initViewportVariables} from './common/viewport';
+import {ESLDemoSidebar} from './navigation/navigation';
+import {ESLDemoMarquee} from './landing/landing';
+
+initViewportVariables();
+
+// Register Demo components
+ESLDemoSidebar.register();
+ESLDemoMarquee.register();
+
+// Register ESL Components
+
 ESLImage.register();
 ESLMedia.register();
 
@@ -61,8 +75,11 @@ ESLA11yGroup.register();
 ESLTabs.register();
 
 ESLScrollbar.register();
+
 ESLAlert.register();
-ESLAlert.init();
+ESLAlert.init({
+  closeOnOutsideAction: true
+});
 
 ESLSelectList.register();
 ESLSelect.register();