+++ src/modules/esl-utils/dom/scroll.ts
@@ -1,3 +1,4 @@
+import {tryUntil} from '../async/promise';
 import {getNodeName, getParentNode} from './api';
 
 export type ScrollStrategy = 'none' | 'native' | 'pseudo';
@@ -102,22 +103,26 @@ export function isScrollParent(element: Element): boolean {
  * This is a promise-based version of scrollIntoView().
  * Method scrolls the element's parent container such that the element on which
  * scrollIntoView() is called is visible to the user. The promise is resolved when
- * the element became visible to the user.
+ * the element became visible to the user and scrolling stops.
+ *
+ * Note: Please, use the native element.scrollIntoView() if you don't need a promise
+ * to detect the moment when the scroll is finished or you don't use smooth behavior.
  * @param element - element to be made visible to the user
  * @param options - scrollIntoView options
  */
-export function scrollIntoViewAsync(element: Element, options: ScrollIntoViewOptions): Promise<void> {
+export function scrollIntoView(element: Element, options?: boolean | ScrollIntoViewOptions | undefined): Promise<boolean> {
+  let same = 0;
+  let lastTop: number;
+  const check = () => {
+    const newTop = element.getBoundingClientRect().top;
+
+    if (newTop !== lastTop) {
+      same = 0;
+      lastTop = newTop;
+    }
+    return same++ > 2;
+  };
+
   element.scrollIntoView(options);
-
-  return new Promise((resolve, reject) => {
-    const intersectionObserver = new IntersectionObserver((entries) => {
-      const [entry] = entries;
-
-      if (entry.isIntersecting) {
-        intersectionObserver.unobserve(element);
-        resolve();
-      }
-    });
-    intersectionObserver.observe(element);
-  });
+  return tryUntil(check, 80, 125); // will check top position every 125ms, but not more than 80 times (10s)
 }