+++ src/modules/esl-utils/misc/format.ts
@@ -1,32 +1,29 @@
 import {get} from './object';
 
-/** Convert string to kebab-case notation */
+/** Converts string to kebab-case notation */
 export const toKebabCase = (str: string) => {
   return str.replace(/([a-z])([A-Z])/g, '$1-$2').replace(/[\s_]+/g, '-').toLowerCase();
 };
 
-/** Convert string to camelCase notation */
+/** Converts string to camelCase notation */
 export const toCamelCase = (str: string) => {
   return str.trim().replace(/[\s-,_]+([a-zA-Z0-9]?)/g, (match: string, word: string) => word.toUpperCase());
 };
 
-/** Make the first non-indent (space, tab, newline) letter in the string capitalized */
+/** Makes the first non-indent (space, tab, newline) letter in the string capitalized */
 export const capitalize = (str: string): string => {
-  for (let i = 0; i < str.length; i++) {
-    if (str.charAt(i) !== ' ') {
-      return str.slice(0, i) + str.charAt(i).toUpperCase() + str.slice(i + 1);
-    }
-  }
-  return str;
+  let i = 0;
+  while (i < str.length && (str[i] === ' ' || str[i] === '\t' || str[i] === '\n')) i++;
+  return str.slice(0, i) + str.charAt(i).toUpperCase() + str.slice(i + 1);
 };
 
-/** Unwrap string from parenthesis */
+/** Unwraps string from parenthesis */
 export const unwrapParenthesis = (str: string) => {
   return str.trim().replace(/^\((.*)\)$/, '$1').trim();
 };
 
 /**
- * Parse number with the ability to pass an alternative fallback for NaN.
+ * Parses number with the ability to pass an alternative fallback for NaN.
  * Note: falsy values except 0 are treated as NaN
  */
 export const parseNumber = (str: string | number, nanValue?: number | undefined) => {
@@ -49,7 +46,7 @@ export function parseAspectRatio(str: string): number {
   return +w || 0;
 }
 
-/** Evaluate passed string or returns `defaultValue` */
+/** Evaluates passed string or returns `defaultValue` */
 export function evaluate(str: string, defaultValue?: any): any {
   try {
     // eslint-disable-next-line @typescript-eslint/no-implied-eval
@@ -60,11 +57,10 @@ export function evaluate(str: string, defaultValue?: any): any {
   }
 }
 
-
 /** Default RegExp to match replacements in the string for the {@link format} function */
 export const DEF_FORMAT_MATCHER = /{[{%]?([\w.]+)[%}]?}/g;
 
-/** Replace `{key}` patterns in the string from the source object */
+/** Replaces `{key}` patterns in the string from the source object */
 export function format(str: string, source: Record<string, any>, matcher: RegExp = DEF_FORMAT_MATCHER) {
   return str.replace(matcher, (match, key) => {
     const val = get(source, key);